Split up each class in its own file, tweaked the evaluate planner class to only work on one planner, added the metric class and sub classes, commented the planner class, and re implemented the straight_line_test2.py
