+++ src/non_lin_opt/ipopt_planner.py
@@ -1,5 +1,8 @@
-from src.utils.classes import *
+from src.utils.planner import Planner
+from src.utils.trajectory_tracking_controller import *
 from src.utils.archive import gif_utils
+import casadi as ca
+import numpy as np
 from src.utils import non_lin_opt_utils, plotting_utils
 import bisect
