+++ src/straight_line/straight_line_planner.py
@@ -1,5 +1,6 @@
-from src.utils.classes import *
+from src.utils.planner import Planner
 import math
+import numpy as np
 
 
 class StraightLinePlanner(Planner):
@@ -14,21 +15,12 @@ class StraightLinePlanner(Planner):
     def __init__(self, problem,
                  settings=None,
                  t_init=None, n=100, mode='open-loop'):
-        Planner.__init__(self, problem, settings, t_init, n, mode)
+        super().__init__(problem, settings, t_init, n, mode)
         # self.dt = self.T_init / self.N
         self.dt = 10.
 
     def get_next_action(self, state):
-        """Go in the direction of the target with full power.
-
-        Args:
-            state:
-                A four element list describing the current state, i.e. [[lon],[lat], [battery_level], [time]]. Note each
-                 nested variable is a float.
-
-        Returns:
-            An array containing the thrust and heading, i.e. array([thrust], [heading]).
-        """
+        """Go in the direction of the target with full power. See superclass for args and return value. """
 
         lon, lat = state[0][0], state[1][0]
         lon_target, lat_target = self.problem.x_T[0], self.problem.x_T[1]
