+++ src/utils/trajectory_tracking_controller.py
@@ -1,0 +1,6 @@
+class TrajectoryTrackingController:
+    """
+    TODO: implement controller to go along planned trajectory for platforms.
+    """
+    def __init(self, trajectory, state):  # Will need planned trajectory and current state
+        pass
